Splitted Controller Tests
